Updated variable type and added test file
Updated variable type to Double for value in items files and added a Item Controller test file.

+++ src/main/java/com/qa/ims/controller/ItemController.java
@@ -46,7 +46,7 @@ public class ItemController implements CrudController<Item>{
 		LOGGER.info("Please enter a product name");
 		String productName = utils.getString();
 		LOGGER.info("Please enter a value");
-		Long value = utils.getLong();
+		Double value = utils.getDouble();
 		Item item = itemDAO.create(new Item(productName, value));
 		LOGGER.info("Item created");
 		return item;
@@ -62,7 +62,7 @@ public class ItemController implements CrudController<Item>{
 		LOGGER.info("Please enter a product name");
 		String productName = utils.getString();
 		LOGGER.info("Please enter a value");
-		Long value = utils.getLong();
+		Double value = utils.getDouble();
 		Item item = itemDAO.update(new Item(id, productName, value));
 		LOGGER.info("Item Updated");
 		return item;
